add `emitDisplayEvents` store action and associated middleware
it was already getting unsustainable to update the display as a direct
store subscriber with just two basic handlers, mainly due to the
difficulty of avoiding a feedback loop by dispatching an action from
the display, which triggers the display's store subscriber, which
causes another action to be dispatched from it, and so on.

by directly emitting display events from the editor we can elide that
problem, since we can easily distinguish where a display event
originated via its `kind` property. for example, the
`editorSelectComponent` event kind corresponds to a selection that
occured within the editor, and in response to this event, the display
triggers a more generic `componentSelect` event encompassing all
component selections, including those originating from the display
itself.